feat: update migration script and mmodel for transactions features

+++ src/main/java/com/hei/wallet/repository/AccountRepository.java
@@ -1,19 +1,19 @@
 package com.hei.wallet.repository;
 
-import com.hei.wallet.models.Account;
-import com.hei.wallet.models.Currency;
+import com.hei.wallet.models.entities.Account;
+import com.hei.wallet.models.entities.Currency;
 
-import java.util.List;
-
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountRepository {
 
-    List<Account> findByBalance(float balance);
+    List<Account> findByBalance(BigDecimal balance);
 
     List<Account> findByCurrencyId(Currency currency);
 
     List<Account> findById(int idReceiver);
 
     List<Account> findByCurrencyId(int currencyId);
+
 }
